requirements.txt
asgiref==3.3.1
astroid==2.4.2
attrs==20.3.0
Automat==20.2.0
cffi==1.14.4
constantly==15.1.0
cryptography==3.3.1
cssselect==1.1.0
Django==3.1.4
djangorestframework==3.12.2
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
isort==5.6.4
itemadapter==0.2.0
itemloaders==1.0.4
jmespath==0.10.0
lazy-object-proxy==1.4.3
lxml==4.6.2
mccabe==0.6.1
numpy==1.19.4
pandas==1.2.0
parsel==1.6.0
Protego==0.1.16
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
PyHamcrest==2.0.2
pylint==2.6.0
pylint-django==2.3.0
pylint-plugin-utils==0.6
pymongo==3.11.2
pyOpenSSL==20.0.1
python-dateutil==2.8.1
pytz==2020.5
queuelib==1.5.0
Scrapy==2.4.1
scrapy-splash==0.7.2
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
Twisted==20.3.0
typed-ast==1.4.1
w3lib==1.22.0
wrapt==1.12.1
zope.interface==5.2.0
